services/crm-core/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.0.post1
alembic==1.13.1
psycopg2-binary==2.9.10
